requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-extra-fields==3.7.0
filetype==1.2.0
pillow==11.2.1
PyJWT==2.9.0
redis==6.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
